% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepel_rstudioapi_2RU.R
\name{deepel_rstudioapi_2RU}
\alias{deepel_rstudioapi_2RU}
\title{Translate Selected Text to Russian via DeepL API}
\usage{
deepel_rstudioapi_2RU()
}
\value{
A message indicating the completion of the translation.
}
\description{
Translate the selected text in RStudio into Russian using the DeepL API with auto-detection of the source language.
   This function uses free-version DeepL API (https://api-free.deepl.com/v2/translate).
}
\details{
Translate Selected Text to Russian via DeepL API

This function uses the rstudioapi to execute Russian translation of the selected text using the DeepL API.
It will attempt to auto-detect the language of the text and translate it into Russian.
First, select the text and execute "DeepL Translation into Russian" from addins.
As a result, the selected part will be translated into Russian.
}
\examples{
\dontrun{

#Select the following text: "La selección está traducida al inglés."
#Then, execute "DeepL Translation into Russian" from RStudio addins.

}
}
\author{
Satoshi Kume
}
