% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepel_rstudioapi_speakInJA.R
\name{deepel_rstudioapi_speakInJA}
\alias{deepel_rstudioapi_speakInJA}
\title{Translate and Speak in Japanese}
\usage{
deepel_rstudioapi_speakInJA()
}
\value{
A message indicating that the translation and speech process has finished.
}
\description{
This function takes the selected text in RStudio, translates it into Japanese using the DeepL API, and then speaks the translated text using MacOS's speech system.
}
\details{
Translate Selected Text and Speak in Japanese via DeepL API and MacOS system
}
\examples{
\dontrun{
#Select the following text: "La selección está traducida al inglés."
#Then, execute "Translation and Speak in Japanese : deepRstudio" from RStudio addins.
}
}
\author{
Satoshi Kume
}
